Make the zero counts dimmer in black and dark theme
The previous PR #12971 made the color of the zero counts in the main screen brighter, in order to make it more visible. However, as a result of this, the contrast between the light gray and the other count colors (blue, red, green) becomes a bit too minimal to differentiate easily at a glance.

This commit will re-adjust the color of the zero. It will make the color dimmer.
The aim of the dimmed zero is to not call attention, so it doesn't so matter if it can't be easily seen.

* Black theme: revert to the color used in 2.15.6
* Dark theme: adjust the color to be a bit lighter than the background, as in the black theme