fix: on Android 13 don't request EXTERNAL_STORAGE
We requested: [`WRITE_EXTERNAL_STORAGE`, `READ_EXTERNAL_STORAGE`]
due to a bug in Android where `WRITE_EXTERNAL_STORAGE` could be
  granted without `READ_EXTERNAL_STORAGE`.

This meant that a user could not access the /AnkiDroid folder
(but it did work from the WebView). This blocked 'Check Media'
and 'Media Sync'.

Re-requesting `READ_EXTERNAL_STORAGE` fixed the permission issue.

Now that we are on `targetSdkVersion 33`, this bug can no longer be
reproduced on Android 13. [See discussion]

As Android 13 also replaces `READ_EXTERNAL_STORAGE` with:
[`READ_MEDIA_IMAGES`, `READ_MEDIA_VIDEO`, `READ_MEDIA_AUDIO`],
these permissions no longer relate, and it makes sense to stop
applying this workaround in case it's still relevant
(as the permissions no longer match the intent)

So: no longer check for these permissions on Android 13

Original bug: 10892

Fixes 14120
Discussion: 14259