Fix selecting the empty default deck in NoteEditor's deck selection
displayDeckSelectionDialog() was calling indirectly decks.allNamesAndIds
allowing for the empty default deck while the computeDropDownDecks()
method used in deck selection called decks.allNamesAndIds with the empty
default deck disalowed.

This resulted in the first deck being used when the default deck was
selected in the dialog.

Commit fixes the issue by changing computeDropDownDecks() to also
include the empty default deck.

Also changed the tests for CardBrowser to handle the default deck being
shown.