Stop using finishWithoutAnimation()
Historically, finishWithoutAnimation() was mostly used as a replacement for finish(), since it was deprecated by previous AnkiDroid devs in favor of using finishWithoutAnimation() and finishWithAnimation().

The issue was that it was common to see contributors, especially new ones (including myself), using finishWithoutAnimation() because only because finish() wasn't available and finishWithAnimation() was harder to use

Since there is no reason for wanting for a screen to finish without animation, which is unnatural, and now that finish() is undeprecated, it should be the default way of closing a screen.

There is no concern about devices with animation, since the behavior will be the same with finish()