Refactor libanki.Storage.getDatabaseVersion()
This inlines the method's content at the only place where it was used,
in CollectionHelper.getDatabaseVersion(). Also, the associated
exception was moved to the anki.exceptions package as it isn't an
exception specific to libanki code.

The code was also improved a bit to use a finally block to close the
database, to avoid situations where the code would crash before the
db.close() call.