Group checks related menu options of DeckPicker into single menu item
This PR combines the three checks actions(check database, check media
and find empty cards) previously available in the options menu of
DeckPicker. The rationale behind this is that these are actions that
are not often used by the user and it would make sense to combine
them to save menu space.

This would make the menu simpler for the average user while also
keeping the options available(at the expense of an extra tap).