Create UI for the new backend backup options in preferences
The logic to handle these preferences was already available through
BackupLimitsPresenter, this PR adds the preferences "UI" and binds it
to BackupLimitsPresenter.

Note that the backup options were set to be observed by analytics,
these can't be used to identify the user and would be useful to
observe to set better defaults.