Enable the backend UI screens for importing an anki package
With the changes in this PR it will only be possible to import a single
 package. Also fixed some minor IDE reported issues.

The progress during the import operation is handled by the backend UI.
Once the native progress dialog was removed we have to call
backend.setWantsAbort() on back button press. This required
implementing the newer predictive back navigation with callbacks that
are enabled/disabled AHEAD of the time of the actual BACK button press.