Implement predictive back navigation for Info
The added back callback is enabled/disabled based on webview status(
can go back or not) to let the system know if it should handle the
back navigation request. The callback is disabled at start, the
WebView it's expected to not have any history yet.

See: https://codelabs.developers.google.com/handling-gesture-back-navigation#5
where this exact scenario is handled.