Handle android dialog cancellation when used as frontend for js dialogs
For js alert types of dialogs this PR adds a cancel listener to call
the expected js api. Without this, if a user cancels the android
dialog in any way besides clicking "ok", the the app will hang.

For js confirm types of dialogs the android dialog was setup as not
cancellable as the user is expected to use one of the provided option,
"ok" or "cancel".