Fix NoteEditor discarding changes on BACK when editing
The deprecated onBackPressed() did the correct checks when user hit
BACK. It also called the super.onBackPressed() which closed the
activity although it was suposed to show an AlertDialog asking the
user for action about his changes.

I replaced it with a OnBackPressedCallback as this doesn't have the
issue with the super call. Note that this callback doesn't handle
the predictive back navigation as it is always on.