feat: match settings app bar and status bar colors
This also changes the `Dark` theme appbar color to match the visual of the other themes.

note: Theme.Dark's colorSurface attribute only had an specific use at the deprecated browser filter sheet dialog, so changing it shouldn't have a relevant impact.

Also note that the color isn't final and will be changed in the future. For now, my goal is to keep things similar to how they looked before so changes can be done incrementally without being too disruptive

That is also the reason of why the status bar color isn't being changed directly in the theme files