Fix ExportReadyDialog not being restored from notification
Implements the DialogHandlerMessage system for the ExportReadyDialog(
the dialog being shown after the collection was exported and which asks
the user what to do with the exported collection). This allows the
dialog to be correctly restored from the notification if the dialog
is being shown while the DeckPicker is not in focus.