NF: more timbers for getting data from http server (#14697)
* NF: more timbers for getting data from http server

I believe that a value not found should be an error. Simply because that's what
one may look for when debugging.

And also, it's important to state when the value is found. Otherwise, when
looking at log, it's not clear that the silence mean the value was actually
found without looking at the codebase.

In the bug I was trying to debug right now, I wanted to check whether it was a
problem with unfound media, or an issue with a media that android does not
render properly. This saved me time.

* change one .e to .w
---------

Co-authored-by: Mike Hardy <github@mikehardy.net>