refactor: use idling resource to stabilize custom scheduler test
Waiting for text to appear on the WebView is not 100% reliable. By using
an idling resource, we can specify exactly when the system is considered
idle and continue with the test. In particular, we wait until the custom
scheduler has persisted its state before clicking on the Show Answer and
Good buttons