test: fix JNI critical lock held error in instrumentation tests
Sometimes the emulator tests are failing because the application gets
forcibly killed by Android. In the logs, we are seeing

    JNI critical lock held for 33.004ms on
    Thread[1,tid=3832,Runnable,Thread*=0x741a9345dbe0,peer=0x71cf55b8,"main"]

which seem to indicate that the Rust backend and database connections to
the collection are leaking across tests and are causing a deadlock. To
fix this, we make sure to discard the Rust backend and close the
collection after each test

Fixes 14940