Enable DeckPicker to handle the selection from DeckPickerContextMenu
The rationale for this change is to have DeckPickerContextMenu as a
standalone dialog fragment and to let DeckPicker do the actual
work(the dialog is just calling methods from the activity). This will
help remove the dependency on the collection for DeckPickerContextMenu
and improve testing.

The only changes are adding a "ContextMenu" prefix for the log calls.