Extract required data for DeckPickerContextMenu as fragment arguments
This removes the need for any collection access in DeckPickerContextMenu as
well the need of a custom factory to pass the collection. The DeckPicker will
have to provide the needed data for the context menu, this makes sense as the
DeckPicker accesses the collection just before showing the menu.

Note: I made the "ensure_cannot_be_instantiated_without_expected_arguments" test
pass just for compatibility, I don't think this test is valuable as any changes
to the context menu without passing the required arguments will crash immediately
when run.