refactor: set which side to play sounds in `play` methods
in the scenario where you want to play the audios of a card, you'd need to do:

```
(show question)
loadCardSounds(FRONT)
playAllSounds()

(flip to answer)
loadCardSounds(BACK)
playAllSounds()
```

that means that the card loading job is done twice, although it is necessary to load the audios only once. So this commit avoids that. The current flow now is like:

```
(show question)
loadCardSounds()
playAllSounds(FRONT)

(flip to answer)
playAllSounds(BACK)
```