start ReadText deprecation process
- AdvancedSettingsFragment:

Show a dialog if the user tries to enable the preference, with a `Learn more` button.

If the user tries to cancel it, the preference will be unchecked as well

- ReadText:

Change the `Initializing...` message to `Please migrate to the new TTS system` with a Learn more link and indefinite time (i.e. the user has to dismiss it manually)