Show toast when importing without the app being started at least once
The current implementation using messages and handlers for async dialogs
means that if the user tries to import before actually starting the app
the import dialog ends up being shown in the introduction screen ending
up in a crash. These changes prevent that by checking for this scenario
and notifying the user that he must start the app at least once before
any import can be done.

The hasShownAppIntro() was moved to better structure the code(it's related
to the introduction screen) and for ease of access in the screens that use it.

(cherry picked from commit 4717ba6d1cff7374bb243035171dc66d88fb2f28)