feat: inform users collection is no longer accessible
After a period of time, Android 13+ will remove
`WRITE_EXTERNAL_STORAGE` if AnkiDroid is unused.

This causes `PermissionsUntil29Fragment` to be shown.

Android does not allow `WRITE_EXTERNAL_STORAGE` to be
re-granted (excluding via `adb`)

So we link a user to the same options as we did in
DIALOG_STORAGE_UNAVAILABLE_AFTER_UNINSTALL

Fixes 14423