fix: remove `object` from parseSourcesToFileScheme
it led to `Same-origin policy` issues with `<svg><object></svg>`.

The only downside is that the loading of `object` will be slower. Since `img`, `video` and `audio` can be used for most cases, uses of `object` should be fairly low so this isn't very relevant

The ideal fix would be to use something like https://github.com/ankidroid/Anki-Android/issues/15863#issuecomment-1992190692, but this should fix it for now.