Feature/add search bar to manage note types (#15789)
* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

* Update AnkiDroid/src/main/java/com/ichi2/anki/notetype/ManageNotetypes.kt

Co-authored-by: David Allison <62114487+david-allison@users.noreply.github.com>

* Delete AnkiDroid/src/main/res/layout/manage_notes_type_toolbar.xml

* Delete AnkiDroid/src/main/res/menu/manage_notes_type_menu.xml

* Delete AnkiDroid/src/main/res/values/strings.xml

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

*Additional Fixes have been made from previous commit such as The manage_notes_type_menu.xml & manage_notes_type_toolbar.xml is removed and standard toolbar and locale_dialog_action_search is used instead , The FirstRun  is been removed*

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

*Additional Fixes have been made from previous commit such as The manage_notes_type_menu.xml & manage_notes_type_toolbar.xml is removed and standard toolbar and locale_dialog_action_search is used instead , The FirstRun  is been removed*

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

*Additional Fixes have been made from previous commit such as The manage_notes_type_menu.xml & manage_notes_type_toolbar.xml is removed and standard toolbar and locale_dialog_action_search is used instead , The FirstRun  is been removed*

* Add search functionality to ManageNotetypes section

This commit introduces a search bar to the ManageNotetypes section in the appbar, enabling users to search for specific note types. The search bar enhances usability and improves the overall user experience.

*Additional Fixes have been made from previous commit such as The manage_notes_type_menu.xml & manage_notes_type_toolbar.xml is removed and standard toolbar and locale_dialog_action_search is used instead , The FirstRun  is been removed*

* *Additional Fixes have been made from previous commit such as background thread using Dispatcher.IO added  ,  null safety checks (?.) added , comments removed  *

* *Additional Fixes have been made from previous commit such as comments are been removed  *

* *Additional Fixes have been made from previous commit such as comments are been removed  *

* *Changes reverted   *

* *Changes reverted   *

* *Changes reverted as follows : toolbar.xml file reverted , additional space in 02-strings.xml removed , space added in NotetypeAdapter.kt *

* *Changes reverted as follows : space added between the methods , searchview?.maxWidth = Integer.MAX_VALUE added , the string name changed to cd_manage_notetypes_add**

* changes made

* error in unit test fixed

---------

Co-authored-by: David Allison <62114487+david-allison@users.noreply.github.com>