Fix displayed name of attahced file when name contains extra points characters
The previous code was blindly splitting after "." which meant that a
filename containing extra "." was displayed just as the before and
after part of the first point chars encountered(discarding everything
else).

The new code will consider the last point encountered of the filename(
which should be the point delimiter for the extension) and split on
that if the filename contains extra points.