Fix #16044: Cannot Scroll to End of TemplateEditor
Keyboard's appearance when editing a card template covers the bottom part of the text, not allowing it to be viewed or to write after it.
Fixed by changing to adjustResize and changing visibility of BottomNavigation bar to only when keyboard is closed, to avoid it covering the visible part above the keyboard.