refactor: Kotlin 2.0 prep
## Deprecations

* `@Deprecated`/ `@Suppress("OVERRIDE_DEPRECATION")`

====

This declaration overrides a deprecated member but is not marked as deprecated itself.
Please add the '@Deprecated' annotation or suppress the diagnostic.
See https://youtrack.jetbrains.com/issue/KT-47902 for details

## `LinkedList.first`

* `LinkedList.first` to `LinkedList.first()`

----

'val first' is deprecated. This declaration will be renamed in a future version of Kotlin.
Please consider using the 'first()' stdlib extension if the collection supports fast random access.

## `kotlinOptions`

* `kotlinOptions` -> `compilerOptions`

 'kotlinOptions(KotlinJvmOptionsDeprecated /* = KotlinJvmOptions */.() -> Unit): Unit' is deprecated.
 Please migrate to the compilerOptions DSL. More details are here: https://kotl.in/u1r8ln