Remove compat code related to notification channels
The compat code for notificationchannels was removed in favor of a
single method that initializes the channels directly if we are in Oreo
and above.

Rationale:
- single use so it doesn't litter the calling code with android version
checks
- although there's an androidx compat alternative that doesn't
implement the lockedScreenVisibility feature that we use for
our notifications.
- the previous compat implementation prevented the move to a
standalone compat module as the compat notification channels
code was reading string resources that were declared in the
main module.