Move compat related code to new compat module
Note: there's an initialization method that was added to CompatHelper's
companion object so we can pass a signaling/reporting callback that can
be used by compat code without the need to depend on AnkiDroid module
classes(it's used in AnkiActivity). This avoids a circular dependency between
the two modules. We should look into if the try-catch code in CustomTabsFallback
is still needed.

Also the VIBRATE permission was moved to the manifest of the new
compat module.