Remove StdModels related code
Used only in tests.

The initializer functions code used to create the models in StdModels
were refactored as standalone functions to be used by interested tests.

Note: StdModels were both used in unit and instrumented tests. As
there's only one usage in an instrumented test I duplicated the code
to avoid the use of the testlib.

The strings for the model names were removed because we don't need
translations for strings used in tests. The "front" and "back" names
are used in CardTemplateEditor, should probably be moved to one of
the other files to not have a strings file with just two strings in
it.