NF: Improve typing of `findViewById`
`findViewById`, being java code without `@Nullable` or `@NotNull`
annotation, returns a value that we sometime consider nullable and
sometime not null. The type checker does not help here.

In API 28 we have `requireViewById` that returns a non null value. We
can't directly use it, but we can copy its definition into
`requireViewByID` and use it. If one day min sdk becomes 28, a
search/replace will allow us to use directly `requireViewById`.

This PR introduce the method and give an example of its usage.