NF: NoteEditor onKeyUp return true
I don't expect it was causing a bug, but `onKeyUp` almost always
returned false, even when the key press was handled.

This is now corrected.

I had trouble understanding the `selectFieldIndex` part. I hope I
improved it. Also added a unit test, given the simplicity of this part
of the code.

I wanted to test the remaining of KeyUtils, but got error messages,
telling me

java.lang.RuntimeException: Method getUnicodeChar in android.view.KeyEvent not mocked. See https://developer.android.com/r/studio-ui/build/not-mocked for details.
	at android.view.KeyEvent.getUnicodeChar(KeyEvent.java)
	at com.ichi2.utils.KeyUtils.getDigit(KeyUtils.kt:30)

I'd be interested in help if it's quick to do. But otherwise, it's not
worth losing time on it.