fix: `Processing...` popup in app startup
8e0b537b1d9fe464a82dad68e8bc744c1cbced4c introduced a call to `updateMenuState()` at the
loadDeckCounts job to update the undo label state when the deck is
reloaded.

Following the method name, the whole menu is updated instead of only the
undo label, which includes checking if there are changes to sync
remotely, and that takes time, so it shouldn't be done if not necessary.