Download crowdin en-translation
This PR would only ensure that the translation tool download the
crowdin en translation. English (GB)

It does not currently use it in any way. It's not even clear how we'll
use it with en-gb from anki, and ensure that people using "english"
keep getting the content of values by default instead of values-en.

So this is really experimental.

Bug: #16820