Ensure the onboarding is displayed after loading the decks
In order to ensure that two views are not simultaneously display, I
had to add a boolean.

I changed the condition, ensuring that not only a deck is displayed,
but it is not empty.

The text tell the user to tap the deck to review it. If the deck is
empty, the text is confusing.