Enhance DeckPicker search icon to include card browser
This adds functionality to search your collection directly from the
deck picker.

The search icon that appears in the deck picker in a collection with
>= 10 decks will currently display a search view for filtering decks.
This change keeps the behavior of filtering decks in the DeckPicker on
typing, but changes the query submit behavior to open a CardBrowser
search for the search query.