Start with size
The problem with current output
> Old APK size: 36215049
> New APK size: 36215063

is that, on any document which is not monospaced (such as github), the
numbers are not aligned. Which means that comparing two numbers is
slightly harder than needed, as digits at the same positions are not
one below the other.

This slight improvement improve the readability.