NF: remove "Delete card" string
This string is misleading, and I'd even state, dangerous.
A card can't be removed. That's a question users often ask, how to
delete a single card. We must NOTE make them believe this is possible.

Using this, if they want to delete a card, they'll end up deleting the
note instead and losing data.

Admittedly, this is not perfect, as the user selected card and not
notes. So it should be "Delete card's note" or "Delete cards' notes"
but that's too long.