Request translator to provide the name of their language
Currently, there is a real issue with the list of languages provided
to the users. In #13275 the list of language was hardcoded, with the
ultimate goal of correcting it eventually.

There are 93 languages. Looking for people speaking those 93 languages
would be very hard. Instead, I suggest we use our standard tool,
crowdin. We'll add an extra file, with a single field to fill, where
the user is invited to put the name of their language, as it should be
displayed in their language.

If this work as expected, we'll have the name of every language
exactly as the translator want it to be displayed, we will NOT have to
guess what is the right string.

In a second time, we'll only have to create a script that update
`AnkiDroid/src/main/java/com/ichi2/utils/LanguageUtil.kt` with the
name provided to us by the translators (or the current name if no name
where provided yet). This script could then either be run with the
update-string script, or once in a while, maybe before releases.