fix(preferences): uppercase language names
The menu provided to the users to select their languages is not yet
perfect. It's hard to know for certain what is the correct way to
display the languages to the users. The function provided by Android
was buggy, and so, a year and a half ago, the language names were
hard-coded in #13275. Except that nobody speaks the 93 languages in
which we have translations available.

On the long term, I hope that 92 translators will provide 92 language
string to use. #17120 should start the process to eventually fix this
error.

On the short term, there is one change that we can make that will
probably be, on average, an improvement. Using upper case for the
first letter of each name.

If I understand correctly Brayan's comment, this change would be
correct for Portuguese. I can confirm it's correct for French. I can't
promise this won't make things worse for some language. But, if we got
some right previously, it was by accident, and I still hope this is,
on average, an improvement.

The upper cases were obtained by using the "set first later to upper
case" feature of emacs on each language name.

íslenska, isiXhosa and isiZulu appear to start with lowercase letters
so these have not been updated

Fixed: #17118

Co-authored-by: David Allison <62114487+david-allison@users.noreply.github.com>