Fix notetype name being reverted when going to CardTemplateEditor
There's a disconnect between what ManageNoteypes is doing(working with
the collection through CollectionManager) and what CardTemplatedEditor
is doing(using the previous system with getColUnsafe). This seems to
result in CardTemplateEditor "seeing" only the previous name from
ManageNotetypes and reverting the name when any change in the template
is saved. A call to Notetypes.clearCache() was introduced to remove
stale data from Notetypes.