Refactor code around ForgetCardsDialog
The refactoring mainly removes the associated ViewModel as it only had
two simple properties and a method to call an activity method to do
its work.

ForgetCardsDialog was refactored:
- simplify the dialog customization
- simplify the dialog itself by removing the need to pass the target
cards ids. With these changes the dialog becomes a simple UI panel
from which the user to select his desired options. The cards ids are
fetched by the activity itself when they are actually needed
- use modern communication(FragmentResult) between the dialog fragment
and the holder activity doing the work
- create utility methods so activities using the dialog can easily
do so